Move main.py's functions into the modules that own them
main.py was 696 lines holding 18 functions. It is now 66 lines that define
exactly one function - main() - listing the run's phases in order.

Into modules that already existed:
  _parse_date_for_sort      -> utils/dates.py
  _crawlbase_token          -> news_fetching/advanced_scraper.py (its
                               original home before the dead-code pass)
  _has_country_partition    -> data_fetching/country_data_fetch.py
  ensure_missing_country_panels -> same, as backfill_missing_panels(),
                               with PROCESSED_DATA travelling along as
                               PANEL_DIR (verified == data_retrieval.DATA_DIR)

New, inside the existing news_fetching/ folder, split pure vs I/O:
  article_ranking.py    - score_relevance, rank_ids_by, ensure_top_three,
                          select_top_ids, impact_topic_maps,
                          build_top_articles. No network; fully unit-tested.
  article_enrichment.py - fetch_relevant_news, resolve_and_enrich,
                          enrich_top_images. All the network work.

New utils/pipeline.py holds the four phases that genuinely span domains
(each calls a fetcher AND an AI ranker AND data_push). Putting them
anywhere else would make data_fetching or ai import data_push and invert
the one-way dependency. main()'s per-country loop moved here too, as
process_all_countries().

Behavior preserved, verified three ways:
- 5 function bodies are AST-identical to their originals; the other 12
  differ only by renamed call sites, extracted named constants, and the
  resilience guards that moved in from main() (each phase now owns its
  own try/except, so the same log messages still fire)
- extracted constants asserted equal to the originals (keyword lists,
  TOP_N=3, threshold 0.3, 24000 chars)
- stdout is unchanged: moved code uses getLogger(__name__) but the
  basicConfig format has no %(name)s

Tests reorganized to mirror the source layout: test_dates.py,
test_article_ranking.py, test_country_data_fetch.py replace
test_main_helpers.py and test_ensure_top_three.py. 185 passed (up from
174 - impact_topic_maps and build_top_articles had no coverage before).
compileall, pyflakes, and the docstring/type-hint audit all clean.
